feat: Enhance ROCm performance and MTP diagnostics
- Updated `test_rocm_check.py` to include GUI path for BuildManager.
- Added MTP recheck documentation (`MTP_RECHECK_2026-05-11.md`) detailing performance metrics and findings on MTP with Qwen3.6 model.
- Created a performance research index (`PERF_RESEARCH_INDEX.md`) for tracking ROCm performance improvements and experiments.
- Documented ROCm kernel performance notes (`PERF_RESEARCH_NOTES.md`) focusing on GDN, FlashAttention, and MMQ optimizations.
- Introduced reduced FlashAttention kernel implementation (`fattn-qwen-reduced.cpp`) to facilitate targeted performance testing.
- Added runtime performance notes (`PERF_RESEARCH_NOTES.md`) to outline key areas affecting Qwen3.6 ROCm performance.

+++ gui/server_tab.py
@@ -536,9 +536,9 @@ class ServerTabWidget(QWidget):
             self._apply_spec_controls_from_extra_args(str(match["extra_args"]).strip())
 
         kv_map = {
-            0: "f32",
-            1: "f16",
-            2: "bf16",
+            0: "f16",
+            1: "bf16",
+            2: "f32",
             3: "q8_0",
             7: "q4_0",
         }
